requirements.txt
anyio==4.9.0
certifi==2025.1.31
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
narwhals==1.34.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
